+++ utilities/app.lua
@@ -1,0 +1,9 @@
+App = Declare('App')
+
+function App:Init()
+	self.parent = Base
+end
+
+function App:Setup()
+	kind('ConsoleApp')
+end